Refactor code formatting and type consistency in sensor-test.go

+++ sensor-test.go
@@ -5,22 +5,21 @@ import (
 	"time"
 )
 
-
-//flow sensor
+// flow sensor
 const (
 	CHIP_NAME string = "gpiochip4"
 	CS_GPIO   int    = 9
 	SPI_DEV   string = "spidev0.0"
 )
 
-//PT sensor
+// PT sensor
 const (
-	PT_SLAVE_ADDR int 			= 0x00
-	MEAS_DELAY 	  time.Duration = 10 * time.Millisecond
-	BUS   		  string     	= "1"
+	PT_SLAVE_ADDR uint8         = 0x00
+	MEAS_DELAY    time.Duration = 10 * time.Millisecond
+	BUS           string        = "1"
 )
 
-func runFlowTest(chipName string, csGPIO int, spiDev string){
+func runFlowTest(chipName string, csGPIO int, spiDev string) {
 	InitSPI(chipName, csGPIO, spiDev)
 	SensorInit()
 	t := time.NewTicker(10 * time.Second)
@@ -30,8 +29,7 @@ func runFlowTest(chipName string, csGPIO int, spiDev string){
 	}
 }
 
-
-func runPtTest(slave_addr int, meas_delay time.Duration, bus string){
+func runPtTest(slave_addr uint8, meas_delay time.Duration, bus string) {
 	s, err := New(bus, slave_addr, meas_delay)
 	if err != nil {
 		panic(err)
@@ -41,7 +39,6 @@ func runPtTest(slave_addr int, meas_delay time.Duration, bus string){
 	fmt.Printf("Temperature: %.2f °C\n", s.GetTemperature())
 	fmt.Printf("Pressure:    %.2f Pa\n", s.GetPressure())
 }
-
 
 func main() {
 	runFlowTest(CHIP_NAME, CS_GPIO, SPI_DEV)
